Split service and gateway
