+++ service/configs/gateway.yaml
@@ -1,0 +1,1 @@
+grpc_gw_port: 8090
